add new file-io tests
add some additional file io tests as well as a script
to run all file-io tests.

+++ src/hip_file_read_all.cc
@@ -37,12 +37,12 @@ static void init_recvbuf (long *recvbuf, int count)
     }
 }
 
-static bool check_recvbuf(long *recvbuf, int nprocs_unused, int rank_unused, int count)
+static bool check_recvbuf(long *recvbuf, int nprocs_unused, int rank, int count)
 {
     bool res=true;
 
     for (long i=0; i<count; i++) {
-        if (recvbuf[i] != i+1) {
+        if (recvbuf[i] != (rank * count) + i+1) {
             res = false;
 #ifdef VERBOSE
             printf("recvbuf[%d] = %ld\n", i, recvbuf[i]);
@@ -53,8 +53,8 @@ static bool check_recvbuf(long *recvbuf, int nprocs_unused, int rank_unused, int
     return res;
 }
 
-int file_read_test (void *sendbuf, int count,
-                    MPI_Datatype datatype, MPI_File fh);
+int file_read_all_test (void *sendbuf, int count,
+                        MPI_Datatype datatype, MPI_File fh);
 
 int main (int argc, char *argv[])
 {
@@ -68,51 +68,60 @@ int main (int argc, char *argv[])
     MPI_Comm_size (MPI_COMM_WORLD, &size);
     MPI_Comm_rank (MPI_COMM_WORLD, &rank);
 
-    if (size > 1){
-        printf("%s only works for 1 process. Aborting\n", argv[0]);
-        MPI_Abort (MPI_COMM_WORLD, 1);
-        return 1;
-    }
-
     parse_args(argc, argv, MPI_COMM_WORLD);
 
     long *tmp_sendbuf=NULL, *tmp_recvbuf=NULL;
     // Initialise send buffer
-    // Forcing the temporary buffer used to write the input data
-    // to be hostbuffer, independent of user input. Only recvbuffer
-    // relevant for the read test.
-    delete sendbuf;
-
-    sendbuf = new hip_mpitest_buffer_host;
-    ALLOCATE_SENDBUFFER(sendbuf, tmp_sendbuf, long, elements, sizeof(long),
-                        rank, MPI_COMM_WORLD, init_sendbuf);
+    if (rank == 0) {
+        // Forcing the temporary buffer used to write the input data
+        // to be hostbuffer, independent of user input. Only recvbuffer
+        // relevant for the read test.
+        delete sendbuf;
+        sendbuf = new hip_mpitest_buffer_host;
+        ALLOCATE_SENDBUFFER(sendbuf, tmp_sendbuf, long, elements * size, sizeof(long),
+                            rank, MPI_COMM_WORLD, init_sendbuf);
+    }
 
     // Initialize recv buffer
     ALLOCATE_RECVBUFFER(recvbuf, tmp_recvbuf, long, elements, sizeof(long),
                         rank, MPI_COMM_WORLD, init_recvbuf);
 
     // Create input file
-    old_mask = umask(022);
-    umask (old_mask);
-    perm = old_mask^0666;
+    if (rank == 0) {
+        old_mask = umask(022);
+        umask (old_mask);
+        perm = old_mask^0666;
 
-    fd = open ("testout.out", O_CREAT|O_WRONLY, perm );
-    if (-1 == fd) {
-        fprintf(stderr, "Error in creating input file. Aborting\n");
-        MPI_Abort (MPI_COMM_WORLD, 1);
-        return 1;
+        fd = open ("testout.out", O_CREAT|O_WRONLY, perm );
+        if (-1 == fd) {
+            fprintf(stderr, "Error in creating input file. Aborting\n");
+            MPI_Abort (MPI_COMM_WORLD, 1);
+            return 1;
+        }
+
+        SL_write(fd, sendbuf->get_buffer(), elements*size*sizeof(long));
+        close (fd);
     }
 
-    SL_write(fd, sendbuf->get_buffer(), elements*sizeof(long));
-    close (fd);
-
     // execute file_read test
-    MPI_File_open(MPI_COMM_SELF, "testout.out", MPI_MODE_RDONLY,
+    MPI_File_open(MPI_COMM_WORLD, "testout.out", MPI_MODE_RDONLY,
                   MPI_INFO_NULL, &fh);
+
+    MPI_Datatype tmptype, fview;
+    MPI_Datatype dtype = MPI_LONG;
+    int blength = elements;
+    MPI_Aint displ = rank * elements * sizeof(long);
+
+    MPI_Type_create_struct (1, &blength, &displ, &dtype, &tmptype);
+    MPI_Type_commit (&tmptype);
+    MPI_Type_create_resized(tmptype, 0, elements*size*sizeof(long), &fview);
+    MPI_Type_commit (&fview);
+    MPI_File_set_view (fh, 0, MPI_LONG, fview, "native", MPI_INFO_NULL);
+
     MPI_Barrier(MPI_COMM_WORLD);
     auto t1s = std::chrono::high_resolution_clock::now();
-    res = file_read_test (recvbuf->get_buffer(), elements,
-                          MPI_LONG, fh);
+    res = file_read_all_test (recvbuf->get_buffer(), elements,
+                              MPI_LONG, fh);
     if (MPI_SUCCESS != res) {
         fprintf(stderr, "Error in file_read_test. Aborting\n");
         MPI_Abort (MPI_COMM_WORLD, 1);
@@ -138,46 +147,27 @@ int main (int argc, char *argv[])
                         elements, (size_t)(elements * sizeof(long)), 1, t1);
 
     //Free buffers
-    FREE_BUFFER(sendbuf, tmp_sendbuf);
-    FREE_BUFFER(recvbuf, tmp_recvbuf);
+    if (rank == 0) {
+        FREE_BUFFER(sendbuf, tmp_sendbuf);
+        delete (sendbuf);
+        unlink("testout.out");
+    }
 
-    delete (sendbuf);
+    FREE_BUFFER(recvbuf, tmp_recvbuf);
     delete (recvbuf);
 
-    unlink("testout.out");
+    MPI_Type_free(&tmptype);
+    MPI_Type_free(&fview);
+
     MPI_Finalize ();
     return fret ? 0 : 1;
 }
 
-int file_read_test (void *recvbuf, int count, MPI_Datatype datatype, MPI_File fh )
+int file_read_all_test (void *recvbuf, int count, MPI_Datatype datatype, MPI_File fh )
 {
     int ret;
 
-#ifdef HIP_MPITEST_FILE_IREAD
-    MPI_Request *req;
-    int i;
-
-    req = (MPI_Request *) malloc (NBLOCKS * sizeof(MPI_Request));
-    if (NULL == req) {
-        fprintf(stderr, "Error allocating memory. Aborting\n");
-        MPI_Abort (MPI_COMM_WORLD, 1);
-        return 1;
-    }
-
-    long *rbuf = (long *)recvbuf;
-    int ncount = count / NBLOCKS;
-    assert ( (count % NBLOCKS) == 0);
-
-    for (i=0; i<NBLOCKS; i++) {
-        ret = MPI_File_iread(fh, rbuf, ncount, datatype, &req[i]);
-        rbuf += ncount;
-    }
-
-    MPI_Waitall (NBLOCKS, req, MPI_STATUS_IGNORE);
-    free (req);
-#else
-    ret = MPI_File_read (fh, recvbuf, count, datatype, MPI_STATUS_IGNORE);
-#endif
+    ret = MPI_File_read_all (fh, recvbuf, count, datatype, MPI_STATUS_IGNORE);
     return ret;
 }
