+++ src/hip_file_write_all.cc
@@ -22,10 +22,10 @@ hip_mpitest_buffer *recvbuf=NULL;
 
 static void SL_read ( int hdl, void *buf, size_t num);
 
-static void init_sendbuf (long *sendbuf, int count, int unused)
+static void init_sendbuf (long *sendbuf, int count, int rank)
 {
     for (long i = 0; i < count; i++) {
-        sendbuf[i] = i+1;
+        sendbuf[i] = (rank*count) + i+1;
     }
 }
 
@@ -36,11 +36,11 @@ static void init_recvbuf (long *recvbuf, int count)
     }
 }
 
-static bool check_recvbuf(long *recvbuf, int nprocs_unused, int rank_unused, int count)
+static bool check_recvbuf(long *recvbuf, int nprocs, int rank_unused, int count)
 {
     bool res=true;
 
-    for (long i=0; i<count; i++) {
+    for (long i=0; i<count*nprocs; i++) {
         if (recvbuf[i] != i+1) {
             res = false;
 #ifdef VERBOSE
@@ -52,8 +52,8 @@ static bool check_recvbuf(long *recvbuf, int nprocs_unused, int rank_unused, int
     return res;
 }
 
-int file_write_test (void *sendbuf, int count,
-                     MPI_Datatype datatype, MPI_File fh);
+int file_write_all_test (void *sendbuf, int count,
+                         MPI_Datatype datatype, MPI_File fh);
 
 int main (int argc, char *argv[])
 {
@@ -67,12 +67,6 @@ int main (int argc, char *argv[])
     MPI_Comm_size (MPI_COMM_WORLD, &size);
     MPI_Comm_rank (MPI_COMM_WORLD, &rank);
 
-    if (size > 1){
-        printf("%s only works for 1 process. Aborting\n", argv[0]);
-        MPI_Abort (MPI_COMM_WORLD, 1);
-        return 1;
-    }
-
     parse_args(argc, argv, MPI_COMM_WORLD);
 
     long *tmp_sendbuf=NULL, *tmp_recvbuf=NULL;
@@ -80,17 +74,32 @@ int main (int argc, char *argv[])
     ALLOCATE_SENDBUFFER(sendbuf, tmp_sendbuf, long, elements, sizeof(long),
                         rank, MPI_COMM_WORLD, init_sendbuf);
 
-    // Initialize recv buffer
-    ALLOCATE_RECVBUFFER(recvbuf, tmp_recvbuf, long, elements, sizeof(long),
-                        rank, MPI_COMM_WORLD, init_recvbuf);
+    if (rank == 0) {
+        // Initialize recv buffer
+        ALLOCATE_RECVBUFFER(recvbuf, tmp_recvbuf, long, elements*size, sizeof(long),
+                            rank, MPI_COMM_WORLD, init_recvbuf);
+    }
+
+    // open file and set file view
+    MPI_File_open(MPI_COMM_WORLD, "testout.out", MPI_MODE_CREATE|MPI_MODE_WRONLY,
+                  MPI_INFO_NULL, &fh);
+
+    MPI_Datatype tmptype, fview;
+    MPI_Datatype dtype = MPI_LONG;
+    int blength = elements;
+    MPI_Aint displ = rank * elements * sizeof(long);
+
+    MPI_Type_create_struct (1, &blength, &displ, &dtype, &tmptype);
+    MPI_Type_commit (&tmptype);
+    MPI_Type_create_resized(tmptype, 0, elements*size*sizeof(long), &fview);
+    MPI_Type_commit (&fview);
+    MPI_File_set_view (fh, 0, MPI_LONG, fview, "native", MPI_INFO_NULL);
 
     // execute file_write test
-    MPI_File_open(MPI_COMM_SELF, "testout.out", MPI_MODE_CREATE|MPI_MODE_WRONLY,
-                  MPI_INFO_NULL, &fh);
     MPI_Barrier(MPI_COMM_WORLD);
     auto t1s = std::chrono::high_resolution_clock::now();
-    res = file_write_test (sendbuf->get_buffer(), elements,
-                           MPI_LONG, fh);
+    res = file_write_all_test (sendbuf->get_buffer(), elements,
+                               MPI_LONG, fh);
     if (MPI_SUCCESS != res) {
         fprintf(stderr, "Error in file_write_test. Aborting\n");
         MPI_Abort (MPI_COMM_WORLD, 1);
@@ -102,10 +111,12 @@ int main (int argc, char *argv[])
 
     // verify results
     bool ret = true;
-    int fd = open ("testout.out", O_RDONLY );
-    if ( -1 != fd ) {
-        SL_read(fd, tmp_recvbuf, elements * sizeof(long));
-        ret = check_recvbuf(tmp_recvbuf, size, rank, elements);
+    if (rank == 0) {
+        int fd = open ("testout.out", O_RDONLY );
+        if ( -1 != fd ) {
+            SL_read(fd, tmp_recvbuf, elements * size * sizeof(long));
+            ret = check_recvbuf(tmp_recvbuf, size, rank, elements);
+        }
     }
 
     bool fret = report_testresult(argv[0], MPI_COMM_WORLD, sendbuf->get_memchar(),
@@ -115,45 +126,26 @@ int main (int argc, char *argv[])
 
     //Free buffers
     FREE_BUFFER(sendbuf, tmp_sendbuf);
-    FREE_BUFFER(recvbuf, tmp_recvbuf);
-
     delete (sendbuf);
-    delete (recvbuf);
 
-    unlink("testout.out");
+    if (rank == 0) {
+        FREE_BUFFER(recvbuf, tmp_recvbuf);
+        delete (recvbuf);
+        unlink("testout.out");
+    }
+
+    MPI_Type_free(&tmptype);
+    MPI_Type_free(&fview);
+
     MPI_Finalize ();
     return fret ? 0 : 1;
 }
 
-int file_write_test (void *sendbuf, int count, MPI_Datatype datatype, MPI_File fh )
+int file_write_all_test (void *sendbuf, int count, MPI_Datatype datatype, MPI_File fh )
 {
     int ret;
 
-#ifdef HIP_MPITEST_FILE_IWRITE
-    MPI_Request *req;
-    int i;
-
-    req = (MPI_Request *) malloc (NBLOCKS * sizeof(MPI_Request));
-    if (NULL == req) {
-        fprintf(stderr, "Error allocating memory. Aborting\n");
-        MPI_Abort (MPI_COMM_WORLD, 1);
-        return 1;
-    }
-
-    long *sbuf = (long *)sendbuf;
-    int ncount = count / NBLOCKS;
-    assert ( (count % NBLOCKS) == 0);
-
-    for (i=0; i<NBLOCKS; i++) {
-        ret = MPI_File_iwrite(fh, sbuf, ncount, datatype, &req[i]);
-        sbuf += ncount;
-    }
-
-    MPI_Waitall (NBLOCKS, req, MPI_STATUS_IGNORE);
-    free (req);
-#else
-    ret = MPI_File_write (fh, sendbuf, count, datatype, MPI_STATUS_IGNORE);
-#endif
+    ret = MPI_File_write_all (fh, sendbuf, count, datatype, MPI_STATUS_IGNORE);
     return ret;
 }
